+++ packages/substrate-client/src/chainhead/call.ts
@@ -3,7 +3,7 @@ import type { OperationCallDoneRpc } from "./json-rpc-types"
 import { createOperationPromise } from "./operation-promise"
 
 export const createCallFn = createOperationPromise(
-  () => chainHead.call,
+  chainHead.call,
   (hash: string, fnName: string, callParameters: string) => [
     [hash, fnName, callParameters],
     (e: OperationCallDoneRpc, res: (output: string) => void) => {